requirements.txt
flask==3.0.0
google-genai==0.3.0
SpeechRecognition>=3.11.0
gTTS==2.5.1
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
static-ffmpeg
